requirements.txt
requests
termcolor
pytesseract
Pillow
atproto
pytumblr
beautifulsoup4
transformers
torch